+++ modular-particle-system/CHANGELOG.md
@@ -12,12 +12,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added `GravityWithCenter` modifier.
 -   Added `RandomColor.interpolate` property.
 -   Added `Circle` shape.
+-   Added `RandomRotation` initializer module.
 
 ### Removed
 
 -   Removed `center` and related properties from `Gravity` modifier.
     -   These were separated to their own module `GravityWithCenter`.
 -   `CircleGenerator` and `CircleExteriorGenerator`, use `ShapeGenerator` instead.
+
+### Fixed
+
+-   `AlphaOverLifetime` not working together with `AlphaRange`
 
 ## [1.1.0] - 2022-04-18